+++ core/src/main/scala/info/mukel/telegrambot4s/methods/EditMessageReplyMarkup.scala
@@ -12,7 +12,7 @@ import info.mukel.telegrambot4s.models.{ChatId, InlineKeyboardMarkup, Message}
   */
 case class EditMessageReplyMarkup(
                                  chatId          : Option[ChatId] = None,
-                                 messageId       : Option[Long] = None,
+                                 messageId       : Option[Int] = None,
                                  inlineMessageId : Option[String] = None,
                                  replyMarkup     : Option[InlineKeyboardMarkup] = None
                                  ) extends ApiRequestJson[Either[Boolean, Message]] {
